backend/requirements.txt
# --- core web ---
fastapi==0.116.1
starlette==0.47.2
uvicorn[standard]==0.35.0

# --- data / json / utils ---
pydantic==2.11.7
pydantic-core==2.33.2
orjson==3.10.6
python-dotenv==1.0.1
regex==2024.11.6

# --- http / async ---
requests==2.32.3
httpx==0.28.1
aiohttp==3.9.5
websockets==15.0.1
anyio==4.4.0

# --- scheduler ---
APScheduler==3.10.4

# --- db (varsa user auth tarafı için) ---
SQLAlchemy==2.0.36
bcrypt==4.2.0
sqlmodel>=0.0.22
python-jose

# --- project-specific (kullandıkların) ---
openai>=1.0,<2.0
tiktoken==0.9.0

google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
